# docs/api/simame-openapi.yaml
openapi: 3.0.3
info:
  title: Simame API
  version: 1.0.0
  description: API documentation for the Simame laundry marketplace.
paths:
...
  /api/v1/orders/{id}/price-breakdown/:
    get:
      operationId: api_v1_orders_price_breakdown_retrieve
      description: Calculates stored totals and applies business logic using FinanceService.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this order.
        required: true
      tags:
      - api
      security:
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderCreate'
          description: ''
...
  /api/v1/orders/items/:
    get:
      operationId: api_v1_orders_items_retrieve
      description: Returns the actual catalog of items with supported services
      tags:
      - api
      security:
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LaunderableItem'
          description: ''
...
components:
  schemas:
...
    LaunderableItem:
      type: object
      description: |-
        ModelSerializer that maps model file/image fields to the safe variants.

        Auto-built fields keep their model-derived kwargs (validators, blank/null),
        but their read path degrades to ``None`` instead of raising when the
        storage backend is misconfigured or unreachable.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          maxLength: 100
        item_category:
          type: string
          format: uuid
          nullable: true
        item_category_name:
          type: string
          readOnly: true
        image:
          type: string
          format: uri
          nullable: true
        is_active:
          type: boolean
      required:
      - id
      - item_category_name
      - name
...
    OrderCreate:
      type: object
      properties:
        laundry:
          type: string
          format: uuid
        pickup_date:
          type: string
          format: date-time
        pickup_address:
          type: string
          nullable: true
        pickup_lat:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,7})?$
          nullable: true
        pickup_lng:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,7})?$
          nullable: true
        delivery_address:
          type: string
          nullable: true
        delivery_lat:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,7})?$
          nullable: true
        delivery_lng:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,7})?$
          nullable: true
        special_instructions:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemCreate'
        payment_method:
          allOf:
          - $ref: '#/components/schemas/OrderCreatePaymentMethodEnum'
          default: CARD
        pricing_mode:
          allOf:
          - $ref: '#/components/schemas/PricingModeEnum'
          default: BY_ITEM
        estimated_weight_kg:
          type: string
          format: decimal
          pattern: ^-?\d{0,4}(?:\.\d{0,2})?$
          nullable: true
      required:
      - laundry
      - pickup_date
    OrderCreatePaymentMethodEnum:
      enum:
      - CARD
      - PAYSTACK
      - CASH
      - CASH_ON_DELIVERY
      - BANK_TRANSFER
      - TRANSFER
      type: string
      description: |-
        * `CARD` - CARD
        * `PAYSTACK` - PAYSTACK
        * `CASH` - CASH
        * `CASH_ON_DELIVERY` - CASH_ON_DELIVERY
        * `BANK_TRANSFER` - BANK_TRANSFER
        * `TRANSFER` - TRANSFER
...
    OrderItemCreate:
      type: object
      properties:
        item:
          type: string
          maxLength: 255
        service_type:
          type: string
          maxLength: 255
        quantity:
          type: integer
          maximum: 99
          minimum: 1
      required:
      - item
      - quantity
      - service_type
...
    PricingModeEnum:
      enum:
      - BY_ITEM
      - BY_WEIGHT
      - CUSTOM_QUOTE
      type: string
      description: |-
        * `BY_ITEM` - By item
        * `BY_WEIGHT` - By weight
        * `CUSTOM_QUOTE` - Pay after quote
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Connect SimpleJWT access token or verified Clerk session JWT.
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid