# ordering/views/order_views.py
# pyre-ignore[missing-module]
from rest_framework import mixins, viewsets, permissions, status
# pyre-ignore[missing-module]
from rest_framework.response import Response
# pyre-ignore[missing-module]
from rest_framework.decorators import action
from rest_framework.exceptions import ValidationError
# pyre-ignore[missing-module]
from ordering.models import LaunderableItem, BookingSlot, Order, Coupon
from ordering.serializers import (
    LaunderableItemSerializer, 
    BookingSlotSerializer, 
    OrderDetailSerializer, 
    OrderCreateSerializer,
    CouponSerializer,
    CouponValidationSerializer
# pyre-ignore[missing-module]
)
# pyre-ignore[missing-module]
from ..services.payment_service import PaymentService
from decimal import Decimal
import logging
import hashlib
from django.core.cache import cache

logger = logging.getLogger(__name__)

class CatalogViewSet(viewsets.ReadOnlyModelViewSet):
    """
    Viewset for global catalog of items and services.
    Strictly filters for active items from approved/active laundries.
    """
    queryset = LaunderableItem.objects.filter(is_active=True).prefetch_related('item_category')
    serializer_class = LaunderableItemSerializer
    permission_classes = [permissions.IsAuthenticated]

    @action(detail=False, methods=['get'])
    def services(self, request):
        """Returns the list of service types (Wash, Iron, etc)"""
# pyre-ignore[missing-module]
        from laundries.models.category import Category
# pyre-ignore[missing-module]
        from laundries.serializers.category import CategorySerializer
        services = Category.objects.filter(type='SERVICE_TYPE')
        serializer = CategorySerializer(services, many=True)
        return Response({
            "status": "success",
            "results": serializer.data
        })

    def list(self, request, *args, **kwargs):
        # Alias for backward compatibility if /booking/services/ was pointing to list
        return self.services(request)

    @action(detail=False, methods=['get'])
    def items(self, request):
        """Returns the actual catalog of items with supported services"""
        queryset = self.get_queryset()
        serializer = self.get_serializer(queryset, many=True)
        return Response({
            "status": "success",
            "results": serializer.data
        })

class BookingViewSet(viewsets.GenericViewSet):
    """Endpoints for booking, scheduling, and creation."""
    permission_classes = [permissions.IsAuthenticated]
    throttle_scope = 'burst_user'
    serializer_class = OrderCreateSerializer

    @action(detail=False, methods=['get'])
    def schedule(self, request):
        laundry_id = request.query_params.get('laundry_id')
        if not laundry_id:
            return Response({"error": "laundry_id is required"}, status=status.HTTP_400_BAD_REQUEST)
        
        slots = BookingSlot.objects.filter(laundry_id=laundry_id, is_available=True)
        serializer = BookingSlotSerializer(slots, many=True)
        return Response({
            "status": "success",
            "message": "Available slots fetched successfully.",
            "data": serializer.data
        })

    @action(detail=False, methods=['post'])
    def estimate(self, request):
        """
        Provides a real-time price breakdown before the user clicks 'Confirm Order'.
        Uses FinanceService and LaundryService for vendor-specific accuracy.
        """
        laundry_id = request.data.get('laundry')
        items_data = request.data.get('items', [])
        
        if not laundry_id or not items_data:
            return Response(
                {"status": "error", "message": "laundry and items are required for estimation."},
                status=status.HTTP_400_BAD_REQUEST
            )
            
        # 1. Create a transient order object (not saved to DB)
        try:
            from laundries.models.laundry import Laundry
            laundry = Laundry.objects.get(id=laundry_id)
        except Laundry.DoesNotExist:
            return Response({"status": "error", "message": "Laundry not found."}, status=status.HTTP_404_NOT_FOUND)

        # 2. Extract specific prices from LaundryService
        from laundries.models.service import LaundryService
        from ordering.models import OrderItem
        
        pickup_lat = request.data.get('pickup_lat')
        pickup_lng = request.data.get('pickup_lng')
        
        # Build a transient order-like object in memory for preview pricing only.
        temp_order = Order(
            laundry=laundry,
            user=request.user,
            pickup_lat=Decimal(str(pickup_lat)) if pickup_lat is not None else None,
            pickup_lng=Decimal(str(pickup_lng)) if pickup_lng is not None else None
        )
        
        total_items_price = Decimal('0.00')
        errors = []
        for data in items_data:
            item_id = data.get('item')
            service_type_id = data.get('service_type')
            quantity = data.get('quantity', 1)
            
            try:
                l_svc = LaundryService.objects.get(laundry_id=laundry_id, item_id=item_id, service_type_id=service_type_id)
                total_items_price += l_svc.price * Decimal(str(quantity))
            except Exception:
                try:
                    from laundries.models.pricing import LaundryPricingItem
                    p_item = LaundryPricingItem.objects.get(laundry_id=laundry_id, id=item_id, is_active=True)
                    total_items_price += p_item.unit_price * Decimal(str(quantity))
                except Exception:
                    errors.append(f"Price not found for item: {item_id}, service: {service_type_id}")
                
        if errors:
            return Response(
                {"status": "error", "message": "Invalid item or service type IDs.", "data": {"errors": errors}},
                status=status.HTTP_400_BAD_REQUEST
            )
                
        # 3. Use FinanceService for the full breakdown
        from ..services.finance_service import FinanceService
        
        outside_service_area = False
        warning_msg = None
        if pickup_lat is not None and pickup_lng is not None:
            lat = float(pickup_lat)
            lng = float(pickup_lng)
            
            if getattr(laundry, 'service_area_polygon', None):
                inside = FinanceService.is_point_in_polygon(lng, lat, laundry.service_area_polygon)
                if not inside:
                    outside_service_area = True
                    warning_msg = "Coordinates are outside the laundry's custom service polygon."
            elif laundry.latitude is not None and laundry.longitude is not None:
                distance = FinanceService.calculate_haversine_distance(
                    lat, lng, laundry.latitude, laundry.longitude
                )
                if distance is not None and distance > float(laundry.service_radius_km):
                    outside_service_area = True
                    warning_msg = f"Coordinates are {distance:.2f} km away, which is outside the laundry's {laundry.service_radius_km} km service radius."

        # Let's do a semi-manual calculation for the preview to avoid DB order creation
        delivery_fee = FinanceService.calculate_delivery_fee(temp_order)
        pickup_fee = FinanceService.calculate_pickup_fee(temp_order)
        # Platform fee & Tax logic. Both are zero-rated by default, so the
        # quote is the laundry's own prices and nothing else.
        tax = FinanceService.calculate_tax_amount(total_items_price)
        platform_fee = FinanceService.calculate_platform_fee(total_items_price)

        total = total_items_price + delivery_fee + pickup_fee + tax + platform_fee

        return Response({
            "status": "success",
            "message": "Price breakdown calculated successfully.",
            "data": {
                "items_total": str(total_items_price.quantize(Decimal('0.01'))),
                "delivery_fee": str(delivery_fee.quantize(Decimal('0.01'))),
                "pickup_fee": str(pickup_fee.quantize(Decimal('0.01'))),
                "tax": str(tax.quantize(Decimal('0.01'))),
                "platform_fee": str(platform_fee.quantize(Decimal('0.01'))),
                "total": str(total.quantize(Decimal('0.01'))),
                "currency": "GHS",
                "delivery_fees_in_app": FinanceService.delivery_fees_in_app(),
                "outside_service_area": outside_service_area,
                "warning": warning_msg
            }
        })

    @action(detail=False, methods=['post'])
    def calculate(self, request):
        """Alias for estimate to match frontend requirement"""
        return self.estimate(request)

    @action(detail=False, methods=['post'])
    def create(self, request):
        idempotency_key = request.headers.get("X-Idempotency-Key")
        cache_key = None
        if idempotency_key:
            body = getattr(getattr(request, '_request', request), 'body', b'') or b''
            request_hash = hashlib.sha256(body).hexdigest()
            fingerprint = hashlib.sha256(
                f"{request.method}:{request.path}:{request_hash}".encode("utf-8")
            ).hexdigest()
            cache_key = f"booking_create_{request.user.id}_{idempotency_key}"
            cached_response = cache.get(cache_key)
            if cached_response:
                if cached_response.get("fingerprint") != fingerprint:
                    return Response(
                        {
                            "status": "error",
                            "message": "This idempotency key was already used for a different booking request.",
                            "data": {},
                        },
                        status=status.HTTP_409_CONFLICT,
                    )
                response = Response(cached_response["data"], status=cached_response["status_code"])
                response["X-Idempotency-Cache"] = "HIT"
                return response

        serializer = OrderCreateSerializer(data=request.data, context={'request': request})
        if serializer.is_valid():
            payment_method = serializer.validated_data.get('payment_method', 'CARD')
            try:
                order = serializer.save()
            except ValidationError as exc:
                return Response(exc.detail, status=status.HTTP_400_BAD_REQUEST)

            # A pay-after-quote order has no price yet, so there is nothing to
            # charge. It waits for the laundry's invoice, which the customer
            # then pays from the order screen through the normal payment path.
            if order.pricing_mode == Order.PricingMode.CUSTOM_QUOTE:
                response_data = OrderDetailSerializer(order).data
                response_data['payment_intent'] = {
                    "transaction_id": None,
                    "amount": "0.00",
                    "currency": "GHS",
                    "status": "QUOTE_PENDING",
                    "payment_method": order.payment_method,
                    "authorization_url": None,
                    "access_code": None,
                    "message": "Pickup requested. You'll get a quote before payment is due.",
                }
                if cache_key:
                    cache.set(
                        cache_key,
                        {"fingerprint": fingerprint, "status_code": status.HTTP_201_CREATED, "data": response_data},
                        86400,
                    )
                return Response(response_data, status=status.HTTP_201_CREATED)

            try:
                payment_info = PaymentService.create_payment_intent(order, payment_method=payment_method)
            except Exception:
                logger.exception(
                    "Payment initialization failed during booking create",
                    extra={"order_id": str(order.id), "request": request},
                )
                payment_info = {
                    "transaction_id": None,
                    "amount": str(order.total_amount),
                    "currency": "GHS",
                    "status": "FAILED",
                    "payment_method": payment_method,
                    "authorization_url": None,
                    "access_code": None,
                    "message": "Order created, but payment could not be initialized. Please retry payment from the receipt.",
                }

            response_data = OrderDetailSerializer(order).data
            response_data['payment_intent'] = payment_info

            if cache_key:
                cache.set(
                    cache_key,
                    {
                        "fingerprint": fingerprint,
                        "status_code": status.HTTP_201_CREATED,
                        "data": response_data,
                    },
                    86400,
                )

            return Response(response_data, status=status.HTTP_201_CREATED)
        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)

class OrderViewSet(mixins.CreateModelMixin, viewsets.ReadOnlyModelViewSet):
    """Viewset for managing and tracking orders."""
    queryset = Order.objects.none()
    permission_classes = [permissions.IsAuthenticated]
    throttle_scope = 'burst_user'

    def get_queryset(self):
        # OrderDetailSerializer reads laundry.name, payment.transaction_reference
        # and (via get_price_breakdown) order.coupon. Without these joins the
        # list endpoint issues three extra queries per order.
        return (
            Order.objects
            .filter(user=self.request.user)
            .select_related('laundry', 'payment', 'coupon')
            .prefetch_related('items')
        )

    def get_serializer_class(self):
        if self.action in ['list', 'retrieve', 'active']:
            return OrderDetailSerializer
        return OrderCreateSerializer

    @action(detail=True, methods=['get'], url_path='price-breakdown')
    def price_breakdown(self, request, pk=None):
        """
        Calculates stored totals and applies business logic using FinanceService.
        """
# pyre-ignore[missing-module]
        from ..services.finance_service import FinanceService
# pyre-ignore[missing-module]
        from django.core.cache import cache

        cache_key = f"order_breakdown_{pk}"
        cached_data = cache.get(cache_key)
        if cached_data:
            return Response({
                "status": "success",
                "message": "Price breakdown fetched (cached)",
                "data": cached_data
            })

        order = self.get_object()
        
        # Security: Only owner or laundry owner
        if order.user != request.user and order.laundry.owner != request.user and not request.user.is_staff:
             return Response({"status": "error", "message": "Unauthorized"}, status=status.HTTP_403_FORBIDDEN)

        # Use centralized finance service
        breakdown = FinanceService.calculate_price_breakdown(order, coupon=order.coupon)

        cache.set(cache_key, breakdown, 300)

        return Response({
            "status": "success",
            "message": "Price breakdown fetched",
            "data": breakdown
        })


    @action(detail=False, methods=['get'])
    def active(self, request):
        """Returns a list of active (in-progress) orders for the current user."""
        active_statuses = [
            Order.Status.PENDING,
            Order.Status.CONFIRMED,
            Order.Status.PICKED_UP,
            Order.Status.IN_PROCESS,
            Order.Status.OUT_FOR_DELIVERY
        ]
        queryset = self.get_queryset().filter(status__in=active_statuses)
        serializer = self.get_serializer(queryset, many=True)
        return Response({
            "status": "success",
            "count": len(serializer.data),
            "results": serializer.data
        })

    @action(detail=True, methods=['get'])
    def tracking(self, request, pk=None):
        """
        Aggregated tracking payload for the customer order detail / tracking screen.

        Returns header info, derived OTP, milestone timeline (one entry per
        backend Order.Status the order has reached or could reach), item
        breakdown, charges/payment summary, laundry info, and the full
        OrderStatusHistory activity log. The schema is stable even when the
        order is brand new — fields that don't apply yet return null/empty.
        """
        from .tracking_view import build_tracking_payload  # local import to avoid cycle
        order = self.get_object()
        payload = build_tracking_payload(order, request=request)
        return Response({
            "status": "success",
            "message": "Order tracking snapshot.",
            "data": payload,
        })

class CouponViewSet(viewsets.GenericViewSet):
    """Viewset for validating and listing available coupons."""
    serializer_class = CouponSerializer
    permission_classes = [permissions.IsAuthenticated]
    
    def get_queryset(self):
        return Coupon.objects.filter(is_active=True)

    @action(detail=False, methods=['post'], url_path='validate')
    def validate(self, request):
        serializer = CouponValidationSerializer(data=request.data)
        serializer.is_valid(raise_exception=True)
        
        code = serializer.validated_data['code']
        laundry_id = serializer.validated_data['laundry_id']
        items_total = serializer.validated_data['items_total']
        
        try:
            coupon = Coupon.objects.get(code=code, is_active=True)
            is_valid, error = coupon.is_valid(
                user=request.user, 
                laundry_id=laundry_id, 
                order_value=items_total
            )
            
            if not is_valid:
                return Response({
                    "status": "error",
                    "message": error,
                    "valid": False
                }, status=status.HTTP_400_BAD_REQUEST)
                
            discount = Decimal('0.00')
            if coupon.discount_type == 'FIXED':
                discount = Decimal(str(coupon.discount_value))
            else:
                discount = (Decimal(str(items_total)) * (Decimal(str(coupon.discount_value)) / 100))
            
            discount = min(discount, Decimal(str(items_total)))
            
            return Response({
                "status": "success",
                "message": "Coupon is valid",
                "valid": True,
                "data": {
                    "code": coupon.code,
                    "discount_amount": str(discount.quantize(Decimal('0.01'))),
                    "discount_type": coupon.discount_type,
                    "discount_value": str(coupon.discount_value)
                }
            })
            
        except Coupon.DoesNotExist:
            return Response({
                "status": "error",
                "message": "Invalid coupon code.",
                "valid": False
            }, status=status.HTTP_404_NOT_FOUND)